git-svn-id: https://gforge.info.ucl.ac.be/svn/mozart@17209 ada56829-ad1f-0410-b00f-83cda6628aec

+++ branches/mozart-gecode/mozart/platform/dss/src/msl_dct.cc
@@ -88,7 +88,7 @@ namespace _msl_internal{
       } else 
 	return false;
     }
-    int len = t_min(static_cast<u32>(bb->availableData()), (a_size - getPosDiff()));
+    int len = t_min(bb->availableData(), (a_size - getPosDiff()));
     bb->readFromBuffer(a_pos, len);
     bb->commitRead(len);
     a_pos += len;
